+++ lib/app/modules/home/sub_views/view_stocks/stock_chart.dart
@@ -4,8 +4,10 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:money/app/controller/preferences_controller.dart';
+import 'package:money/app/controller/selection_controller.dart';
 import 'package:money/app/core/helpers/chart_helper.dart';
 import 'package:money/app/core/helpers/date_helper.dart';
+import 'package:money/app/core/helpers/misc_helpers.dart';
 import 'package:money/app/core/helpers/string_helper.dart';
 import 'package:money/app/core/widgets/center_message.dart';
 import 'package:money/app/core/widgets/chart.dart';
@@ -14,9 +16,22 @@ import 'package:money/app/core/widgets/gaps.dart';
 import 'package:money/app/core/widgets/icon_button.dart';
 import 'package:money/app/data/storage/get_stock_from_cache_or_backend.dart';
 
-class StockChartWidget extends StatefulWidget {
-  const StockChartWidget({required this.symbol, super.key});
+class HoldingActivity {
+  HoldingActivity(this.date, this.amount, this.quantity);
 
+  final double amount;
+  final DateTime date;
+  final double quantity;
+}
+
+class StockChartWidget extends StatefulWidget {
+  const StockChartWidget({
+    super.key,
+    required this.symbol,
+    required this.holdingsActivities,
+  });
+
+  final List<HoldingActivity> holdingsActivities;
   final String symbol;
 
   @override
@@ -30,7 +45,7 @@ class StockChartWidgetState extends State<StockChartWidget> {
   @override
   void initState() {
     super.initState();
-    getStockHistoricalData();
+    _getStockHistoricalData();
   }
 
   @override
@@ -58,24 +73,60 @@ class StockChartWidgetState extends State<StockChartWidget> {
     switch (latestPriceHistoryData.status) {
       case StockLookupStatus.foundInCache:
       case StockLookupStatus.validSymbol:
-        return _buildChart();
       case StockLookupStatus.invalidSymbol:
-        return Center(child: Text('Symbol "${latestPriceHistoryData.symbol.toUpperCase()}" not found.'));
+        return _buildChart();
       default:
         return const Center(child: Text('loading...'));
     }
   }
 
-  void getStockHistoricalData() async {
+  void _getStockHistoricalData() async {
     StockPriceHistoryCache priceCache = await getFromCacheOrBackend(widget.symbol);
 
     _fromPriceHistoryToChartDataPoints(priceCache);
   }
 
+  void _adjustMissingDataPointInthePast() {
+    for (final activiy in widget.holdingsActivities.reversed) {
+      if (dataPoints.isEmpty || activiy.date.millisecondsSinceEpoch < dataPoints.first.x) {
+        dataPoints.insert(
+          0,
+          FlSpot(
+            activiy.date.millisecondsSinceEpoch.toDouble(),
+            activiy.amount,
+          ),
+        );
+      }
+    }
+  }
+
   Widget _buildChart() {
+    const double marginLeft = 80;
+    const double marginBottom = 50;
+
+    // Date ascending
+    dataPoints.sort((a, b) => a.x.compareTo(b.x));
+
+    _adjustMissingDataPointInthePast();
+
+    if (dataPoints.isEmpty) {
+      return const CenterMessage(message: 'No data points');
+    }
+
     return Stack(
       alignment: Alignment.topCenter,
       children: [
+        Padding(
+          padding: const EdgeInsets.only(left: marginLeft, bottom: marginBottom),
+          child: CustomPaint(
+            size: const Size(double.infinity, double.infinity),
+            painter: VerticalBarsPainter(
+              activities: widget.holdingsActivities,
+              minX: dataPoints.first.x,
+              maxX: dataPoints.last.x,
+            ),
+          ),
+        ),
         LineChart(
           LineChartData(
             lineBarsData: [
@@ -88,21 +139,21 @@ class StockChartWidgetState extends State<StockChartWidget> {
               leftTitles: const AxisTitles(
                 sideTitles: SideTitles(
                   showTitles: true,
-                  reservedSize: 80,
+                  reservedSize: marginLeft,
                   getTitlesWidget: getWidgetChartAmount,
                 ),
               ),
               bottomTitles: AxisTitles(
                 sideTitles: SideTitles(
                   showTitles: true,
-                  reservedSize: 50,
+                  reservedSize: marginBottom,
                   getTitlesWidget: (double value, TitleMeta meta) {
                     if (value == meta.min || value == meta.max) {
                       return const SizedBox();
                     }
                     final date = DateTime.fromMillisecondsSinceEpoch(value.toInt());
                     return Text(
-                      formatDate(date),
+                      DateFormat('yyyy\nMMM').format(date),
                       textAlign: TextAlign.center,
                       style: const TextStyle(fontSize: 10),
                     ); // Format as HH:MM
@@ -117,7 +168,7 @@ class StockChartWidgetState extends State<StockChartWidget> {
                 getTooltipItems: (List<LineBarSpot> touchedSpots) {
                   return touchedSpots.map((touchedSpot) {
                     return LineTooltipItem(
-                      doubleToCurrency(touchedSpot.y),
+                      '${doubleToCurrency(touchedSpot.y)}\n${dateToString(DateTime.fromMillisecondsSinceEpoch(touchedSpot.x.toInt()))}\n${getActivityQuantity(touchedSpot.x.toInt())}',
                       const TextStyle(color: Colors.white),
                     );
                   }).toList();
@@ -131,6 +182,15 @@ class StockChartWidgetState extends State<StockChartWidget> {
         _buildPriceRefreshButton(),
       ],
     );
+  }
+
+  String getActivityQuantity(final int fromMillisecondsSinceEpoch) {
+    final activityFound =
+        widget.holdingsActivities.firstWhereOrNull((a) => a.date.millisecondsSinceEpoch == fromMillisecondsSinceEpoch);
+    if (activityFound == null) {
+      return '';
+    }
+    return '${getIntAsText(activityFound.quantity.toInt())} shares';
   }
 
   Widget _buildPriceRefreshButton() {
@@ -187,6 +247,43 @@ class StockChartWidgetState extends State<StockChartWidget> {
   }
 }
 
-String formatDate(DateTime date) {
-  return DateFormat('yyyy\nMMM').format(date);
+class VerticalBarsPainter extends CustomPainter {
+  VerticalBarsPainter({
+    required this.activities,
+    required this.minX,
+    required this.maxX,
+  });
+
+  final List<HoldingActivity> activities;
+  final double maxX;
+  final double minX;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final paintSharesBought = Paint()
+      ..color = Colors.orange.withOpacity(0.8)
+      ..style = PaintingStyle.fill;
+
+    final paintSharesSold = Paint()
+      ..color = Colors.blue.withOpacity(0.8)
+      ..style = PaintingStyle.fill;
+
+    final chartWidth = size.width;
+    final chartHeight = size.height;
+
+    for (var activity in activities) {
+      double left = 0;
+      debugLog('$activity $minX');
+      if (activity.date.millisecondsSinceEpoch > minX) {
+        left = ((activity.date.millisecondsSinceEpoch - minX) / (maxX - minX)) * chartWidth;
+      }
+      final rect = Rect.fromLTWH(left, 0, 1, chartHeight);
+      canvas.drawRect(rect, activity.quantity > 0 ? paintSharesBought : paintSharesSold);
+    }
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return false;
+  }
 }